study_center/requirements.txt
apiview==1.3.25
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-admin-interface==0.17.0
django-colorfield==0.4.3
django-cors-headers==3.8.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.6.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.8
permission==0.4.1
Pillow==8.3.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
teacher==0.0.2
xlrd==2.0.1
xlwt==1.3.0
